소속 ContentView, Screen 분리
